Split request logic into custom Hooks, Fix All Lint
