+++ lib/useUserScores.js
@@ -1,43 +1,48 @@
-//
-//  WARNING: THIS FILE SHOULD BE TESTED.
-//
-import useSWR from 'swr'
+
 import { getToken } from './userAuth'
+import { useState, useEffect } from 'react'
 
 export const USER_SCORES_URL = `${process.env.NEXT_PUBLIC_API_URL}/user_scores`
 
-const useUserScores = UserId => {
+export default function useUserScores(UserId) {
 
-  const fetcher = async url => {
-    const res = await fetch(url, {
-      method: 'POST',
-      headers: {
-        Authorization: `Bearer ${getToken()}`,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        user_id: UserId,
-      }),
-    })
+  const [ scores, setScores ] = useState([ ])
+  const [ error, setMessage ] = useState('')
 
-    if (!res.ok) {
-      const error = new Error('An error occured while fetching the data.')
-      // Attach extra info to the error object.
-      error.info = await res.json()
-      error.status = res.status
-      throw error
+  // Set the scores array of a user. If something goes wrong, the "error" state
+  // is set.
+  useEffect(() => {
+    if (!UserId) {
+      return
     }
 
-    return res.json().then(data => data.scores)
-  }
+    // Request the scores of a specific user.
+    const fetchData = async () => {
+      let response = await fetch(USER_SCORES_URL, {
+        method: 'POST',
+        headers: {
+          Authorization: `Bearer ${getToken()}`,
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          user_id: UserId,
+        }),
+      })
+      response = await response.json()
+      if (response.errors) {
+        setMessage(response.errors)
+      } else {
+        setScores(response.scores)
+        setMessage(undefined)
+      }
+    }
 
+    // Effect Logic
+    fetchData().catch(e => alert(e))
+  }, [ UserId ])
 
-  const { data, error } = useSWR(USER_SCORES_URL, fetcher)
   return {
-    scores: data,
-    error: error && error.message,
+    scores,
+    error,
   }
 }
-
-
-export default useUserScores
